@@ -71,51 +83,90 @@
         "d3-selection": "1.4.1",
         "d3-shape": "^1.3.7",
         "d3-zoom": "^1.8.3",
-        "hashids": "2.1.0",
+        "dayjs": "^1.8.24",
+        "hashids": "2.2.1",
         "js-cookie": "^2.2.1",
-        "lit-html": "^1.1.2",
+        "lit-html": "^1.2.1",
         "moment": "^2.24.0",
-        "round-slider": "^1.3.3",
-        "semver": "^7.1.2",
+        "round-slider": "^1.4.1",
+        "semver": "^7.3.0",
         "tablesort": "^5.1.0",
         "tempusdominus-bootstrap-4": "^5.1.2",
         "tempusdominus-core": "^5.0.3",
-        "timelines-chart": "^2.8.4"
+        "timelines-chart": "^2.9.0"
     },
     "devDependencies": {
-        "@babel/core": "^7.7.5",
+        "@babel/cli": "^7.8.4",
+        "@babel/core": "^7.9.0",
+        "@babel/plugin-proposal-class-properties": "^7.8.3",
         "@babel/plugin-proposal-nullish-coalescing-operator": "^7.8.3",
         "@babel/plugin-transform-spread": "^7.7.4",
-        "@babel/preset-env": "^7.7.5",
+        "@babel/preset-env": "^7.9.5",
+        "@babel/preset-typescript": "^7.9.0",
+        "@microsoft/tsdoc": "^0.12.19",
+        "@types/bootstrap": "^4.3.2",
+        "@types/copy-webpack-plugin": "^5.0.0",
+        "@types/css": "^0.0.31",
+        "@types/d3": "^5.7.2",
+        "@types/d3-array": "^2.0.0",
+        "@types/d3-axis": "^1.0.12",
+        "@types/d3-collection": "^1.0.8",
+        "@types/d3-drag": "^1.2.3",
+        "@types/d3-format": "^1.3.1",
+        "@types/d3-scale": "^2.2.0",
+        "@types/d3-selection": "^1.4.1",
+        "@types/d3-shape": "^1.3.2",
+        "@types/d3-zoom": "^1.7.4",
+        "@types/google.analytics": "^0.0.40",
+        "@types/html-minifier": "^3.5.3",
+        "@types/html-minifier-terser": "^5.0.0",
+        "@types/jquery": "^3.3.34",
+        "@types/js-cookie": "^2.2.6",
+        "@types/lzma-native": "^4.0.0",
+        "@types/node-fetch": "^2.5.6",
+        "@types/node-sass": "^4.11.0",
+        "@types/pngjs": "^3.4.2",
+        "@types/polylabel": "^1.0.5",
+        "@types/recursive-readdir": "^2.2.0",
+        "@types/sharp": "^0.24.0",
+        "@types/terser-webpack-plugin": "^2.2.0",
+        "@types/twit": "^2.2.23",
+        "@types/webpack": "^4.41.11",
+        "@types/webpack-dev-server": "^3.10.1",
+        "@types/webpack-env": "^1.15.1",
+        "@types/webpack-subresource-integrity": "^1.2.0",
+        "@typescript-eslint/eslint-plugin": "^2.28.0",
+        "@typescript-eslint/parser": "^2.28.0",
         "angular-html-parser": "^1.3.0",
-        "autoprefixer": "^9.7.3",
-        "babel-eslint": "^10.0.3",
-        "babel-loader": "^8.0.6",
+        "autoprefixer": "^9.7.6",
+        "babel-eslint": "^10.1.0",
+        "babel-loader": "^8.1.0",
         "clean-webpack-plugin": "^3.0.0",
         "copy-webpack-plugin": "^5.1.0",
-        "core-js": "^3.4.7",
+        "core-js": "^3.6.5",
+        "cross-env": "^7.0.2",
         "css": "^2.2.4",
-        "css-loader": "^3.3.0",
-        "d3-node": "^2.2.1",
-        "dayjs": "^1.8.19",
+        "css-loader": "^3.5.2",
         "denque": "^1.4.1",
-        "diff2html-cli": "^4.2.1",
-        "eslint": "^6.7.2",
-        "eslint-config-prettier": "^6.7.0",
-        "eslint-config-xo": "^0.27.2",
-        "eslint-plugin-import": "^2.18.2",
+        "diff2html-cli": "^5.0.4",
+        "eslint": "^6.8.0",
+        "eslint-config-prettier": "^6.10.1",
+        "eslint-config-xo": "^0.29.1",
+        "eslint-config-xo-typescript": "^0.27.0",
         "eslint-plugin-lit": "^1.2.0",
-        "eslint-plugin-prettier": "^3.1.1",
-        "eslint-plugin-unicorn": "^16.0.0",
+        "eslint-plugin-prettier": "^3.1.3",
+        "eslint-plugin-tsdoc": "^0.2.4",
+        "eslint-plugin-unicorn": "^18.0.1",
         "excel4node": "^1.7.2",
         "exports-loader": "^0.7.0",
         "extract-css-chunks-webpack-plugin": "^4.7.1",
         "favicons-webpack-plugin": "^3.0.1",
         "fibers": "^4.0.2",
-        "file-loader": "^5.0.2",
+        "file-loader": "^6.0.0",
+        "fork-ts-checker-webpack-plugin": "^4.1.3",
         "git": "^0.1.5",
-        "html-webpack-plugin": "^4.0.0-beta.11",
-        "husky": "^4.2.1",
+        "html-webpack-plugin": "^4.2.0",
+        "husky": "^4.2.5",
         "hyper": "^0.0.9",
         "image-webpack-loader": "^6.0.0",
         "imports-loader": "^0.8.0",
@@ -127,56 +178,59 @@
         "node-fetch": "^2.6.0",
         "node-sass": "^4.13.0",
         "npm-run-all": "^4.1.5",
-        "object-merge-advanced": "^10.11.8",
-        "pngjs": "^3.4.0",
+        "object-merge-advanced": "^10.11.16",
+        "pngjs": "^4.0.0",
         "polylabel": "^1.0.2",
         "postcss": "^7.0.24",
         "postcss-clean": "^1.1.0",
         "postcss-cli": "^7.1.0",
         "postcss-loader": "^3.0.0",
         "postcss-scss": "^2.0.0",
-        "prettier": "^1.19.1",
-        "prettier-plugin-packagejson": "^2.0.10",
+        "prettier": "^2.0.4",
+        "prettier-plugin-packagejson": "^2.2.0",
         "prettier-stylelint": "^0.4.2",
         "recursive-readdir": "^2.2.2",
-        "remark-parse": "^7.0.2",
-        "sass": "^1.25.0",
+        "remark-parse": "^8.0.0",
+        "sass": "^1.26.3",
         "sass-loader": "^8.0.0",
         "sassdoc": "^2.7.1",
         "shapefile": "^0.6.6",
-        "sharp": "^0.24.0",
-        "shellcheck": "^0.4.1",
-        "sitemap-webpack-plugin": "^0.6.0",
-        "string-replace-loader": "^2.2.0",
-        "stylelint": "^13.0.0",
+        "sharp": "^0.25.2",
+        "shellcheck": "^0.4.2",
+        "sitemap-webpack-plugin": "^0.7.1",
+        "string-replace-loader": "^2.3.0",
+        "stylelint": "^13.3.2",
         "stylelint-config-prettier": "^8.0.1",
         "stylelint-config-twbs-bootstrap": "^2.0.0",
         "stylelint-prettier": "^1.1.2",
-        "svg-url-loader": "^3.0.3",
+        "svg-url-loader": "^5.0.0",
         "terser-webpack-plugin": "^2.2.2",
         "twit": "^2.2.11",
-        "url-loader": "^3.0.0",
-        "webpack": "^4.41.2",
-        "webpack-bundle-analyzer": "^3.5.2",
+        "typescript": "^3.8.3",
+        "url-loader": "^4.1.0",
+        "utility-types": "^3.10.0",
+        "webpack": "^4.42.1",
+        "webpack-bundle-analyzer": "^3.7.0",
         "webpack-cli": "^3.3.9",
         "webpack-dev-server": "^3.8.1",
         "webpack-manifest-plugin": "^2.2.0",
         "webpack-mode": "^1.0.2",
         "webpack-subresource-integrity": "^1.3.3",
-        "workbox-webpack-plugin": "^5.0.0",
+        "workbox-webpack-plugin": "^5.1.2",
         "xml-js": "^1.6.11",
         "xss": "^1.0.6",
-        "yaml": "^1.7.0",
+        "yaml": "^1.8.3",
         "yaml-unist-parser": "^1.0.0",
-        "yarn-deduplicate": "^1.1.1"
+        "yarn-deduplicate": "^2.0.0"
     },
     "peerDependencies": {
+        "@types/node": "^13.9.1",
         "jquery": "^3.4.1",
         "moment-timezone": "^0.5.14",
         "popper.js": "^1.16.0"
     },
     "engines": {
-        "node": ">=12"
+        "node": ">=13"
     },
     "sass": "src/scss/main.scss"
 }
